Reset server attempt count after each commit checkout
This is the number of times we have tried to connect to the server
after checking out the next commit and the setup script has been
run. It should therefore be a 'per commit' count, rather than
a global count for the whole process.

The main reason behind this is that sometimes a server can take
a while to render the first page. For example Rails can take a few
seconds to precompile all the assets when the source code changes.